move more mip stuff to mip module
